refactor(scalars): move AIDInput to ui
