+++ packages/design-system/src/scalars/components/aid-field/aid-field.tsx
@@ -1,168 +1,9 @@
-/* eslint-disable react/jsx-props-no-spreading */
-import React, { useCallback, useId, useMemo } from "react";
 import { isAddress } from "viem";
-import { IdAutocompleteContext } from "../fragments/id-autocomplete/id-autocomplete-context.js";
-import { IdAutocompleteListOption } from "../fragments/id-autocomplete/id-autocomplete-list-option.js";
-import { IdAutocomplete } from "../fragments/id-autocomplete/index.js";
+import { AIDInput } from "../../../ui/components/data-entry/aid-input/index.js";
 import { withFieldValidation } from "../fragments/with-field-validation/index.js";
-import type { FieldErrorHandling, InputBaseProps } from "../types.js";
-import type { AIDOption, AIDProps } from "./types.js";
+import type { AIDFieldProps } from "./types.js";
 
-type AIDFieldBaseProps = Omit<
-  React.InputHTMLAttributes<HTMLInputElement>,
-  | keyof InputBaseProps<string>
-  | keyof FieldErrorHandling
-  | keyof AIDProps
-  | "pattern"
->;
-
-export type AIDFieldProps = AIDFieldBaseProps &
-  InputBaseProps<string> &
-  FieldErrorHandling &
-  AIDProps;
-
-const AIDFieldRaw = React.forwardRef<HTMLInputElement, AIDFieldProps>(
-  (
-    {
-      id: idProp,
-      name,
-      className,
-      label,
-      description,
-      value,
-      defaultValue,
-      disabled,
-      placeholder,
-      required,
-      errors,
-      warnings,
-      onChange,
-      onBlur,
-      onClick,
-      onMouseDown,
-      supportedNetworks,
-      autoComplete: autoCompleteProp,
-      variant = "withValue",
-      maxLength,
-      fetchOptionsCallback,
-      fetchSelectedOptionCallback,
-      isOpenByDefault, // to be used only in stories
-      initialOptions, // to be used only in stories
-      previewPlaceholder,
-      ...props
-    },
-    ref,
-  ) => {
-    const prefix = useId();
-    const id = idProp ?? `${prefix}-aid`;
-    const autoComplete = autoCompleteProp ?? true;
-
-    const contextValue = useMemo(
-      () => ({ supportedNetworks }),
-      [supportedNetworks],
-    );
-
-    const renderOption = useCallback(
-      (
-        option: AIDOption,
-        displayProps?: {
-          asPlaceholder?: boolean;
-          showValue?: boolean;
-          isLoadingSelectedOption?: boolean;
-          handleFetchSelectedOption?: (value: string) => void;
-          isFetchSelectedOptionSync?: boolean;
-          className?: string;
-        },
-      ) => (
-        <IdAutocompleteListOption
-          variant={variant}
-          icon={option.icon}
-          title={option.title}
-          path={
-            displayProps?.asPlaceholder
-              ? previewPlaceholder?.path || "URL not available"
-              : option.path
-          }
-          value={
-            displayProps?.asPlaceholder
-              ? previewPlaceholder?.value || "aid not available"
-              : option.value
-          }
-          description={option.description}
-          agentType={
-            displayProps?.asPlaceholder
-              ? previewPlaceholder?.agentType || "Agent type not available"
-              : option.agentType
-          }
-          placeholderIcon={previewPlaceholder?.icon || "Person"}
-          {...displayProps}
-        />
-      ),
-      [variant, previewPlaceholder],
-    );
-
-    return (
-      <IdAutocompleteContext.Provider value={contextValue}>
-        {autoComplete && fetchOptionsCallback ? (
-          <IdAutocomplete
-            id={id}
-            name={name}
-            className={className}
-            label={label}
-            description={description}
-            value={value}
-            defaultValue={defaultValue}
-            disabled={disabled}
-            placeholder={placeholder}
-            required={required}
-            errors={errors}
-            warnings={warnings}
-            onChange={onChange}
-            onBlur={onBlur}
-            onClick={onClick}
-            onMouseDown={onMouseDown}
-            autoComplete={true}
-            variant={variant}
-            maxLength={maxLength}
-            fetchOptionsCallback={fetchOptionsCallback}
-            fetchSelectedOptionCallback={fetchSelectedOptionCallback}
-            isOpenByDefault={isOpenByDefault}
-            initialOptions={initialOptions}
-            renderOption={renderOption}
-            previewPlaceholder={previewPlaceholder}
-            {...props}
-            ref={ref}
-          />
-        ) : (
-          <IdAutocomplete
-            id={id}
-            name={name}
-            className={className}
-            label={label}
-            description={description}
-            value={value}
-            defaultValue={defaultValue}
-            disabled={disabled}
-            placeholder={placeholder}
-            required={required}
-            errors={errors}
-            warnings={warnings}
-            onChange={onChange}
-            onBlur={onBlur}
-            onClick={onClick}
-            onMouseDown={onMouseDown}
-            autoComplete={false}
-            maxLength={maxLength}
-            {...props}
-            ref={ref}
-          />
-        )}
-      </IdAutocompleteContext.Provider>
-    );
-  },
-);
-
-export const AIDField = withFieldValidation<AIDFieldProps>(AIDFieldRaw, {
+const AIDField = withFieldValidation<AIDFieldProps>(AIDInput, {
   validations: {
     _validAIDFormat:
       ({ supportedNetworks }) =>
@@ -220,3 +61,5 @@ export const AIDField = withFieldValidation<AIDFieldProps>(AIDFieldRaw, {
 });
 
 AIDField.displayName = "AIDField";
+
+export { AIDField };
