+++ packages/design-system/src/scalars/components/aid-field/index.ts
@@ -1,1 +1,2 @@
-export * from "./aid-field.js";
+export { AIDField } from "./aid-field.js";
+export type { AIDFieldProps } from "./types.js";
