+++ packages/design-system/src/scalars/components/aid-field/types.ts
@@ -1,45 +1,6 @@
-import type {
-  IdAutocompleteOption,
-  IdAutocompleteProps,
-} from "../fragments/id-autocomplete/types.js";
+import type { AIDInputProps } from "../../../ui/components/data-entry/aid-input/types.js";
+import type { FieldErrorHandling } from "../types.js";
 
-export interface Network {
-  chainId: string;
-  name?: string;
-}
+type AIDFieldProps = AIDInputProps & FieldErrorHandling;
 
-type AIDOptionProps = {
-  agentType?: string;
-};
-export type AIDOption = IdAutocompleteOption<AIDOptionProps>;
-
-type AIDBaseProps = Omit<
-  IdAutocompleteProps,
-  | "autoComplete"
-  | "fetchOptionsCallback"
-  | "fetchSelectedOptionCallback"
-  | "renderOption"
-> & {
-  supportedNetworks?: Network[];
-};
-
-export type AIDProps = AIDBaseProps &
-  (
-    | {
-        autoComplete: false;
-        fetchOptionsCallback?: never;
-        fetchSelectedOptionCallback?: never;
-        previewPlaceholder?: never;
-      }
-    | {
-        autoComplete?: true;
-        fetchOptionsCallback: (
-          userInput: string,
-          context?: Record<string, unknown>,
-        ) => Promise<AIDOption[]> | AIDOption[];
-        fetchSelectedOptionCallback?: (
-          value: string,
-        ) => Promise<AIDOption | undefined> | AIDOption | undefined;
-        previewPlaceholder?: AIDOption;
-      }
-  );
+export type { AIDFieldProps };
