+++ packages/design-system/src/ui/components/data-entry/aid-input/index.ts
@@ -1,0 +1,2 @@
+export { AIDInput } from "./aid-input.js";
+export type { AIDInputProps, AIDOption } from "./types.js";
